Mostly just moved stuff around.
Added this paragraph

Before you get started, let me note that there is a lot of complexity here.
All of it is needed.  It is the distilled wisdon of many people working on this
for more than [20 years](https://en.wikipedia.org/wiki/Zope#History).
Even I, the author of this document, had to go back and read it multiple times,
each time going one layer deeper in understanding what is going on.  So be
patient with yourself.